rename ClerkPetitionDataDisplay to ClerkDataDisplay
