+++ src/views/ClerkView.vue
@@ -2,7 +2,7 @@
   <v-container fluid>
     <v-row>
       <v-col cols="12">
-        <ClerkPetitionDataDisplay
+        <ClerkDataDisplay
           :petition="selectedPetition"
           @close="selectPetition(null)"
           @refresh="handleRefresh"
@@ -32,7 +32,7 @@ import { computed, onMounted, onUnmounted, ref, watch } from 'vue';
 import { useStore } from 'vuex';
 import { useI18n } from 'vue-i18n';
 import ContentApiService from '@/services/contentApiService';
-import ClerkPetitionDataDisplay from '@/components/clerk/ClerkPetitionDataDisplay.vue';
+import ClerkDataDisplay from '@/components/clerk/ClerkDataDisplay.vue';
 import ClerkPetitionTable from '@/components/tables/ClerkPetitionTable.vue';
 import loginErrorHandler from '@/utils/loginErrorHandler';
 import { log } from '@/utils/log';
